Enable auto-title for non-storyStoreV7 (#218)
Fixes #201 

This enables the omission of `title` properties in the default export of stories (the new storybook "autotitle" feature).  It's a little less powerful without https://github.com/eirslett/storybook-builder-vite/pull/182, but it's still nice to support this, I think.

I did notice an issue with HMR, but it seems not to be introduced in this PR, but exists in the react example on main as well.  Changing a css value _does_ still HMR, so I think there is no regression here.