Test built storybooks in CI (#274)
So far, we've only been building our storybooks, but not doing any testing on them.  This adds https://github.com/storybookjs/test-runner which will render each of the stories and smoke test them (make sure that there are no runtime errors).  This is at least a step towards having more confidence that our stories aren't broken.  The next step would be to actually write assertions in the stories to be sure they're rendering _correctly_, but that can be done another time.