Move import of framework module after configs (#289)
This should fix https://github.com/storybookjs/storybook/issues/17773.  You can check it in the vue example of this repo.  In main, the docs tab does not work, and neither does the introduction story.  But with this PR, it does.  This also aligns us more closely with the webpack builder, which imports the `configure` function in the virtual story entry, not the config entry.  I tested this change in our examples back to storybook 6.4.0.