Check for explicit typescript dep for react-docgen (#313)
Fixes #312.

Previously, we'd just look to see if typescript was in the project at all, but turns out that it can be included as a peer dependency of some storybook dependencies.

So, this tightens the check a bit, to search the `package.json` for `typescript` in either `devDependencies` or `dependencies` (in case someone forgot to use `--save-dev` I guess).  

To verify, I removed the `typescript` config from our react example (that should have been a clue to me this was broken, I assumed it was due to our monorepo), and I also checked that `vite:react-docgen-typescript` is still being used in our react-ts example.