Support React 18 (#320)
An [exciting PR](https://github.com/storybookjs/storybook/pull/17215) has been merged and released in `6.5.0-alpha.58 ` that adds storybook support for React 18.  However, it currently breaks in the vite-builder test because of an `import()` of the new `react-dom/client`, which only exists in react 18.  Due to https://github.com/vitejs/vite/issues/6007, we have to do a little bit of trickery to stop vite from erroring out.  The approach I'm taking here is:

1) Rewriting the import of `react-dom/client` to a dummy file if it can't be `require.resolve()`ed.  This fixes the problem in production builds.
2) Additionally, adding `react-dom/client` to the list of `optimizeDeps.exclude` if the framework is react and the package can't be required.  This is necessary to prevent vite from trying to pre-bundle, and throwing errors in dev.

This PR also updates the examples to the new `6.5.0-alpha.58` version of storybook, which includes the PR mentioned above, so that it can be tested in our react examples, and it adds a new `react-18` example as well.