Avoid @fs imports for storyStoreV7 (#332)
Similar to https://github.com/storybookjs/builder-vite/pull/324, this avoids the use of `/@fs/` imports in the modern loader, with `storyStoreV7` enabled.

To test, use the react-ts example, install `@storybook/addon-bench`, add it to the `addons` list in `main.js`, and try starting storybook.  It will fail in main, but pass in this branch.