Organize stories in examples (#330)
This adds basic READMEs to all the examples (we can expand/improve them as needed), to try to give some information about why we include these examples. Since some are used to demonstrate features other than just the framework (e.g. autotitle or storyStoreV7).

This also organizes stories a bit in the auto-title examples, and adds some comments in the stories about why they do or don't have titles.

I think that our examples can serve not only as tests for ourselves, but education for users getting started with storybook and the vite builder, so I'd love to make them as useful as possible for both purposes.