react-docgen: exclude node_modules, make configurable (#338)
This updates the internal react-docgen plugin to prevent running against files that are in node_modules, and makes it align with rollup plugin best practices by accepting include and exclude options.

I realized we should do this when working on converting a large project over to use the vite builder as an experiment, and found that it took ~20 minutes to load the first storybook page after the browser opened.  After digging, I found that most of this time was spent processing files that were in other parts of the monorepo, like icons, which did not need docgen info.  After making the changes in this PR, the time was reduced to less than 2 minutes on a cold start.  The number of processed files dropped from 1668 to 958.  And the time drops to 20 seconds for a cold boot with storyStoreV7 (processes 39 files).

This is configurable, although it's not super-straightforward.  For example:

```js
async viteFinal(config) {
    config.plugins = [
      require('@storybook/builder-vite/dist/plugins/react-docgen').reactDocgen({
        exclude: [/^\.\.\//, /node_modules/],
        include: /\.(jsx?)$/,
      }),
      ...config.plugins.filter((plugin) => plugin?.name !== 'react-docgen'),
    ];
    return config;
}
```

Note that `react-docgen` needs to happen before the `@vitejs/plugin-react` plugins, which is why it's above the spread of the rest of the plugins in this example.