Do not require @storybook/addon-svelte-csf (#388)
Starting in Storybook 6.5, `@storybook/addon-svelte-csf` is not installed by default in Svelte projects.  Instead, it's encouraged to use CSF for stories.

But, the vite builder expects that it's always there, and would throw an error if it's not installed.  This allows it to be missing.

It also fixes the logic for our other `catch` statements, which weren't working because the boolean logic was reversed.  🤦.