Add webpack/cra migration section to readme (#405)
This adds a bit of help for users who are coming from a webpack or create react app project. There are a few gotchas and sneaky things that are needed, which we should work to smooth out, but for now teaching users how to migrate seems like it might be useful.