Ignore applyDecorators in v6 story store (#443)
The applyDecorators is an export from frameworks (and can be overridden in preview.js) but is only used in storyStoreV7 mode. The file being updated here is used for the v6 store, which handles decorators a different way, but still ends up reading the export which falls through to the default warning. We can safely ignore it in the case statement.