Support vite 3.1 (#490)
I think we actually can support it just fine now, but our build fails due to a change in the Typescript definitions of rollup hooks.  Here's an example of a failing vite ecosystem-ci build: https://github.com/vitejs/vite-ecosystem-ci/runs/8107408224?check_suite_focus=true

So, this updates the examples to vite 3.1.0-beta.1, adds a `devDependency` on that version for the builder itself, and does a check before calling the react-refresh `transform` function to make typescript happy.