Move examples out of workspaces (#502)
We sometimes have trouble with our examples because of hoisting issues. So, instead of using a monorepo, this PR would turn this repo into a multi-project repo, with no top-level package.json, and instead each example would be stand-alone, and so would packages/builder-vite.

I'm not positive this is the way to go, but it did already suss out a few issues with missing dependencies in both the projects as well as the builder-vite itself.

One thing that's a bit wonky is that since the builder-vite has a completely different node_modules, it can't use any of the packages that are installed in the example, like preact and the vite preact plugin in the preact example. I added them as dev deps to the builder, which works fine I think. It's a downside, but on the other hand it makes these deps a little bit explicit without bloating the dependencies for users.