Fix vue 2.x selection logic (#508)
Fixes: https://github.com/storybookjs/builder-vite/issues/506

This corrects the vue 2.6 / 2.7 logic, and also finds a slightly-better way to handle module resolution from within `builder-vite` when running examples.  Instead of adding dev dependencies into the builder, this sets a `node_path` environment variable, so that the node_modules of the example is used as a global lookup from within the builder.