Update @vitejs/plugin-react dependency for vite 4 (#548)
Fixes https://github.com/storybookjs/builder-vite/issues/547

This also updates the version of mdx1-csf to one that no longer requires the use of the react vite plugin to compile jsx, it handles compilation itself.  So, we no longer need to include `@vitejs/plugin-react` in all projects, and react projects will need to make sure it's installed.  That makes this a breaking change.

And this also updates Vite to 4 in the examples, and fixes some missing dependencies.

---------

Co-authored-by: Josh Wooding <12938082+joshwooding@users.noreply.github.com>